Move ViewModel to correct package. Correct package of ContactReasonScreen.kt
