+++ androidApp/src/main/java/com/EENX15_22_17/digital_journal/android/data/ViewModelModule.kt
@@ -2,7 +2,7 @@ package com.EENX15_22_17.digital_journal.android.data
 
 import com.EENX15_22_17.digital_journal.android.screens.journal.JournalEntryViewModel
 import com.EENX15_22_17.digital_journal.android.screens.journalList.JournalListViewModel
-import com.EENX15_22_17.digital_journal.android.ui.screen.ContactReasonViewModel
+import com.EENX15_22_17.digital_journal.android.screens.journal.triage.contactReason.ContactReasonViewModel
 import org.kodein.di.DI
 import org.kodein.di.bindMultiton
 import org.kodein.di.bindSingleton
