+++ androidApp/src/main/java/com/EENX15_22_17/digital_journal/android/screens/journal/triage/contactReason/ContactReasonViewModel.kt
@@ -1,4 +1,4 @@
-package com.EENX15_22_17.digital_journal.android.ui.screen
+package com.EENX15_22_17.digital_journal.android.screens.journal.triage.contactReason
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
